+++ src/api/pet.js
@@ -1,8 +1,8 @@
-import instance from './axios';
+import { instance, localInstance } from './axios';
 
 export const memberData = async () => {
   try {
-    const response = await instance.get('/api/member');
+    const response = await localInstance.get('/api/member');
     return response.data;
   } catch (error) {
     console.error('Error 확인:', error);
